app, test: create separate *Action objects
