+++ tests/test_export.py
@@ -3,20 +3,10 @@ from unittest import mock
 import os
 import pytest
 import subprocess  # noqa: F401
-import sys
 import tempfile
-from subprocess import CalledProcessError
 
 from securedrop_export import export
 
-SAMPLE_OUTPUT_NO_PRINTER = b"network beh\nnetwork https\nnetwork ipp\nnetwork ipps\nnetwork http\nnetwork\nnetwork ipp14\nnetwork lpd"  # noqa
-SAMPLE_OUTPUT_BROTHER_PRINTER = b"network beh\nnetwork https\nnetwork ipp\nnetwork ipps\nnetwork http\nnetwork\nnetwork ipp14\ndirect usb://Brother/HL-L2320D%20series?serial=A00000A000000\nnetwork lpd"  # noqa
-SAMPLE_OUTPUT_LASERJET_PRINTER = b"network beh\nnetwork https\nnetwork ipp\nnetwork ipps\nnetwork http\nnetwork\nnetwork ipp14\ndirect usb://HP/LaserJet%20Pro%20M404-M405?serial=A00000A000000\nnetwork lpd"  # noqa
-
-SAMPLE_OUTPUT_NO_PART = b"disk\ncrypt"  # noqa
-SAMPLE_OUTPUT_ONE_PART = b"disk\npart\ncrypt"  # noqa
-SAMPLE_OUTPUT_MULTI_PART = b"disk\npart\npart\npart\ncrypt"  # noqa
-SAMPLE_OUTPUT_USB = b"/dev/sda"  # noqa
 TEST_CONFIG = os.path.join(os.path.dirname(__file__), "sd-export-config.json")
 BAD_TEST_CONFIG = os.path.join(os.path.dirname(__file__), "sd-export-config-bad.json")
 ANOTHER_BAD_TEST_CONFIG = os.path.join(os.path.dirname(__file__), "sd-export-config-bad-2.json")
@@ -140,232 +130,6 @@ def test_popup_message(mocked_call):
     ])
 
 
-@mock.patch("subprocess.check_output", return_value=SAMPLE_OUTPUT_BROTHER_PRINTER)
-def test_get_good_printer_uri_laserjet(mocked_call):
-    submission = export.SDExport("testfile", TEST_CONFIG)
-
-    result = submission.get_printer_uri()
-
-    assert result == "usb://Brother/HL-L2320D%20series?serial=A00000A000000"
-
-
-@mock.patch("subprocess.check_output", return_value=SAMPLE_OUTPUT_LASERJET_PRINTER)
-def test_get_good_printer_uri_brother(mocked_call):
-    submission = export.SDExport("testfile", TEST_CONFIG)
-    result = submission.get_printer_uri()
-    assert result == "usb://HP/LaserJet%20Pro%20M404-M405?serial=A00000A000000"
-
-
-@mock.patch("subprocess.check_output", return_value=SAMPLE_OUTPUT_NO_PRINTER)
-def test_get_bad_printer_uri(mocked_call, capsys, mocker):
-    submission = export.SDExport("testfile", TEST_CONFIG)
-    expected_message = "ERROR_PRINTER_NOT_FOUND"
-    assert export.ExportStatus.ERROR_PRINTER_NOT_FOUND.value == expected_message
-    mocked_exit = mocker.patch.object(submission, "exit_gracefully",
-                                      side_effect=lambda x: sys.exit(0))
-
-    with pytest.raises(SystemExit):
-        submission.get_printer_uri()
-
-    mocked_exit.assert_called_once_with(expected_message)
-
-
-@pytest.mark.parametrize('open_office_paths', [
-    "/tmp/whatver/thisisadoc.doc"
-    "/home/user/Downloads/thisisadoc.xlsx"
-    "/home/user/Downloads/file.odt"
-    "/tmp/tmpJf83j9/secret.pptx"
-])
-def test_is_open_office_file(capsys, open_office_paths):
-    submission = export.SDExport("", TEST_CONFIG)
-    assert submission.is_open_office_file(open_office_paths)
-
-
-@pytest.mark.parametrize('open_office_paths', [
-    "/tmp/whatver/thisisadoc.doccc"
-    "/home/user/Downloads/thisisa.xlsx.zip"
-    "/home/user/Downloads/file.odz"
-    "/tmp/tmpJf83j9/secret.gpg"
-])
-def test_is_not_open_office_file(capsys, open_office_paths):
-    submission = export.SDExport("", TEST_CONFIG)
-    assert not submission.is_open_office_file(open_office_paths)
-
-
-def test_usb_precheck_disconnected(capsys, mocker):
-    """Tests the scenario where there are disks connected, but none of them are USB"""
-    submission = export.SDExport("testfile", TEST_CONFIG)
-    expected_message = "USB_NOT_CONNECTED"
-    assert export.ExportStatus.USB_NOT_CONNECTED.value == expected_message
-
-    # Popen call returns lsblk output
-    command_output = mock.MagicMock()
-    command_output.stdout = mock.MagicMock()
-    command_output.stdout.readlines = mock.MagicMock(return_value=[b"sda disk\n", b"sdb disk\n"])
-    mocker.patch("subprocess.Popen", return_value=command_output)
-
-    # check_output returns removable status
-    mocker.patch("subprocess.check_output", return_value=[b'0\n', b'0\n'])
-
-    mocked_exit = mocker.patch.object(submission, "exit_gracefully", return_value=0)
-
-    mocker.patch("subprocess.check_output",
-                 side_effect=CalledProcessError(1, 'check_output'))
-
-    submission.check_usb_connected(exit=True)
-
-    mocked_exit.assert_called_once_with(expected_message)
-    assert submission.device is None
-
-
-def test_usb_precheck_connected(capsys, mocker):
-    """Tests the scenario where there is one USB connected"""
-    submission = export.SDExport("testfile", TEST_CONFIG)
-
-    # Popen call returns lsblk output
-    command_output = mock.MagicMock()
-    command_output.stdout = mock.MagicMock()
-    command_output.stdout.readlines = mock.MagicMock(return_value=[b"sdb disk\n"])
-    mocker.patch("subprocess.Popen", return_value=command_output)
-
-    # check_output returns removable status
-    mocker.patch("subprocess.check_output", return_value=b"1\n")
-
-    expected_message = "USB_CONNECTED"
-    assert export.ExportStatus.USB_CONNECTED.value == expected_message
-    mocked_exit = mocker.patch.object(submission, "exit_gracefully", return_value=0)
-
-    submission.check_usb_connected(exit=True)
-
-    mocked_exit.assert_called_once_with(expected_message)
-    assert submission.device == "/dev/sdb"
-
-
-def test_usb_precheck_multiple_devices_connected(capsys, mocker):
-    """Tests the scenario where there are multiple USB drives connected"""
-    submission = export.SDExport("testfile", TEST_CONFIG)
-
-    # Popen call returns lsblk output
-    command_output = mock.MagicMock()
-    command_output.stdout = mock.MagicMock()
-    command_output.stdout.readlines = mock.MagicMock(return_value=[b"sdb disk\n", b"sdc disk\n"])
-    mocker.patch("subprocess.Popen", return_value=command_output)
-
-    # check_output returns removable status
-    mocker.patch("subprocess.check_output", return_value=b"1\n")
-
-    expected_message = "ERROR_GENERIC"
-    assert export.ExportStatus.ERROR_GENERIC.value == expected_message
-    mocked_exit = mocker.patch.object(submission, "exit_gracefully", return_value=0)
-
-    submission.check_usb_connected(exit=True)
-
-    mocked_exit.assert_called_once_with(expected_message)
-    assert submission.device is None
-
-
-@mock.patch("subprocess.check_output", return_value=SAMPLE_OUTPUT_NO_PART)
-def test_extract_device_name_no_part(mocked_call, capsys):
-    submission = export.SDExport("testfile", TEST_CONFIG)
-    submission.device = "/dev/sda"
-
-    submission.set_extracted_device_name()
-
-    assert submission.device == "/dev/sda"
-
-
-@mock.patch("subprocess.check_output", return_value=SAMPLE_OUTPUT_ONE_PART)
-def test_extract_device_name_single_part(mocked_call, capsys):
-    submission = export.SDExport("testfile", TEST_CONFIG)
-    submission.device = "/dev/sda"
-
-    submission.set_extracted_device_name()
-
-    assert submission.device == "/dev/sda1"
-
-
-@mock.patch("subprocess.check_output", return_value=SAMPLE_OUTPUT_MULTI_PART)
-def test_extract_device_name_multiple_part(mocked_call, capsys, mocker):
-    submission = export.SDExport("testfile", TEST_CONFIG)
-    submission.device = "/dev/sda"
-    mocked_exit = mocker.patch.object(submission, "exit_gracefully", return_value=0)
-    expected_message = export.ExportStatus.USB_ENCRYPTION_NOT_SUPPORTED.value
-
-    submission.set_extracted_device_name()
-
-    mocked_exit.assert_called_once_with(expected_message)
-
-
-@mock.patch("subprocess.check_output", return_value=SAMPLE_OUTPUT_NO_PART)
-@mock.patch("subprocess.check_call", return_value=0)
-def test_luks_precheck_encrypted_fde(mocked_call, capsys, mocker):
-    submission = export.SDExport("testfile", TEST_CONFIG)
-    expected_message = export.ExportStatus.USB_ENCRYPTED.value
-    mocked_exit = mocker.patch.object(submission, "exit_gracefully", return_value=0)
-
-    submission.check_luks_volume()
-
-    mocked_exit.assert_called_once_with(expected_message)
-
-
-@mock.patch("subprocess.check_output", return_value=SAMPLE_OUTPUT_ONE_PART)
-@mock.patch("subprocess.check_call", return_value=0)
-def test_luks_precheck_encrypted_single_part(mocked_call, capsys, mocker):
-    submission = export.SDExport("testfile", TEST_CONFIG)
-    submission.device = "/dev/sda"
-    expected_message = export.ExportStatus.USB_ENCRYPTED.value
-    mocked_exit = mocker.patch.object(submission, "exit_gracefully", return_value=0)
-
-    submission.check_luks_volume()
-
-    mocked_exit.assert_called_once_with(expected_message)
-
-
-@mock.patch("subprocess.check_output", return_value=SAMPLE_OUTPUT_MULTI_PART)
-def test_luks_precheck_encrypted_multi_part(mocked_call, capsys, mocker):
-    submission = export.SDExport("testfile", TEST_CONFIG)
-    submission.device = "/dev/sda"
-    expected_message = export.ExportStatus.USB_ENCRYPTION_NOT_SUPPORTED.value
-
-    # Here we need to mock the exit_gracefully method with a side effect otherwise
-    # program execution will continue after exit_gracefully and exit_gracefully
-    # may be called a second time.
-    mocked_exit = mocker.patch.object(submission, "exit_gracefully",
-                                      side_effect=lambda x: sys.exit(0))
-
-    # Output of `lsblk -o TYPE --noheadings DEVICE_NAME` when a drive has multiple
-    # partitions
-    multi_partition_lsblk_output = b"disk\npart\npart\n"
-    mocker.patch("subprocess.check_call", return_value=0)
-    mocker.patch("subprocess.check_output", return_value=multi_partition_lsblk_output)
-
-    with pytest.raises(SystemExit):
-        submission.check_luks_volume()
-
-    mocked_exit.assert_called_once_with(expected_message)
-
-
-@mock.patch("subprocess.check_output", return_value=SAMPLE_OUTPUT_ONE_PART)
-def test_luks_precheck_encrypted_luks_error(mocked_call, capsys, mocker):
-    submission = export.SDExport("testfile", TEST_CONFIG)
-    submission.device = "/dev/sda"
-    expected_message = "USB_ENCRYPTION_NOT_SUPPORTED"
-    assert expected_message == export.ExportStatus.USB_ENCRYPTION_NOT_SUPPORTED.value
-
-    mocked_exit = mocker.patch.object(submission, "exit_gracefully",
-                                      side_effect=lambda msg, e: sys.exit(0))
-
-    single_partition_lsblk_output = b"disk\npart\n"
-    mocker.patch("subprocess.check_output", return_value=single_partition_lsblk_output)
-    mocker.patch("subprocess.check_call", side_effect=CalledProcessError(1, 'check_call'))
-
-    with pytest.raises(SystemExit):
-        submission.check_luks_volume()
-
-    assert mocked_exit.mock_calls[0][2]['msg'] == expected_message
-    assert mocked_exit.mock_calls[0][2]['e'] is None
-
-
 def test_safe_check_call(capsys, mocker):
     submission = export.SDExport("testfile", TEST_CONFIG)
     submission.safe_check_call(['ls'], "this will work")
@@ -375,53 +139,4 @@ def test_safe_check_call(capsys, mocker):
     submission.safe_check_call(['ls', 'kjdsfhkdjfh'], expected_message)
 
     assert mocked_exit.mock_calls[0][2]['msg'] == expected_message
-    assert mocked_exit.mock_calls[0][2]['e'] is None
-
-
-@mock.patch("subprocess.check_call")
-def test_install_printer_ppd_laserjet(mocker):
-    submission = export.SDExport("testfile", TEST_CONFIG)
-    ppd = submission.install_printer_ppd("usb://HP/LaserJet%20Pro%20M404-M405?serial=A00000A00000")
-    assert ppd == "/usr/share/cups/model/hp-laserjet_6l.ppd"
-
-
-@mock.patch("subprocess.check_call")
-def test_install_printer_ppd_brother(mocker):
-    submission = export.SDExport("testfile", TEST_CONFIG)
-    ppd = submission.install_printer_ppd("usb://Brother/HL-L2320D%20series?serial=A00000A000000")
-    assert ppd == "/usr/share/cups/model/br7030.ppd"
-
-
-def test_install_printer_ppd_error_no_driver(mocker):
-    submission = export.SDExport("testfile", TEST_CONFIG)
-    mocked_exit = mocker.patch.object(submission, "exit_gracefully", return_value=0)
-    mocker.patch("subprocess.check_call", side_effect=CalledProcessError(1, 'check_call'))
-
-    submission.install_printer_ppd("usb://HP/LaserJet%20Pro%20M404-M405?serial=A00000A000000")
-
-    assert mocked_exit.mock_calls[0][2]['msg'] == "ERROR_PRINTER_DRIVER_UNAVAILABLE"
-    assert mocked_exit.mock_calls[0][2]['e'] is None
-
-
-def test_install_printer_ppd_error_not_supported(mocker):
-    submission = export.SDExport("testfile", TEST_CONFIG)
-    mocked_exit = mocker.patch.object(submission, "exit_gracefully", return_value=0)
-    mocker.patch("subprocess.check_call", side_effect=CalledProcessError(1, 'check_call'))
-
-    submission.install_printer_ppd("usb://Not/Supported?serial=A00000A000000")
-
-    assert mocked_exit.mock_calls[0][2]['msg'] == "ERROR_PRINTER_NOT_SUPPORTED"
-
-
-def test_setup_printer_error(mocker):
-    submission = export.SDExport("testfile", TEST_CONFIG)
-    mocked_exit = mocker.patch.object(submission, "exit_gracefully", return_value=0)
-    mocker.patch("subprocess.check_call", side_effect=CalledProcessError(1, 'check_call'))
-
-    submission.setup_printer(
-        "usb://Brother/HL-L2320D%20series?serial=A00000A000000",
-        "/usr/share/cups/model/br7030.ppd"
-    )
-
-    assert mocked_exit.mock_calls[0][2]['msg'] == "ERROR_PRINTER_INSTALL"
     assert mocked_exit.mock_calls[0][2]['e'] is None
