+++ tests/test_main.py
@@ -1,0 +1,3 @@
+from securedrop_export.main import __main__  # noqa: F401
+# This import ensures at least the imports in main.__main__
+# are executed during a test run
